fix: preserve imessage action lazy boundary
Move the iMessage outbound allowlist assertion into a lightweight policy
module so action registration no longer imports the full send runtime.
Keep the previous send.ts export available for internal compatibility.

Verification:
- git diff --cached --check
- node scripts/run-vitest.mjs extensions/imessage/src/send.test.ts extensions/imessage/src/actions.test.ts extensions/imessage/src/channel.runtime.test.ts extensions/imessage/src/monitor/deliver.test.ts extensions/imessage/src/test-plugin.test.ts extensions/imessage/src/monitor.gating.test.ts
- CI=1 PNPM_CONFIG_VERIFY_DEPS_BEFORE_RUN=false corepack pnpm check:changed

Note: the documented Blacksmith Testbox check was rejected by sandbox policy
because it would export unpublished staged code to an external runner.
