+++ extensions/imessage/src/actions.ts
@@ -24,8 +24,8 @@ import {
   rememberIMessageReplyCache,
   type IMessageChatContext,
 } from "./monitor-reply-cache.js";
+import { assertIMessageOutboundAllowed } from "./outbound-allowlist.js";
 import { getCachedIMessagePrivateApiStatus, probeIMessagePrivateApi } from "./probe.js";
-import { assertIMessageOutboundAllowed } from "./send.js";
 import { parseIMessageTarget, type IMessageTarget } from "./targets.js";
 
 const loadIMessageActionsRuntime = createLazyRuntimeNamedExport(
